pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "cloudy-intell"
version = "0.1.0"
description = "Multi-agent cloud architecture generation and validation workflow"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Niranjan Gade" }
]
dependencies = [
    "langchain>=1.2.10",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=1.0.10",
    "langsmith>=0.7.14",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
cloudy-intell = "cloudy_intell.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cloudy_intell"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
exclude = ["Development/"]
